refactor: split src into packages
Move `src/react` into `packages/use-fireproof` and the rest from `src`
into `packages/core`.

This makes `use-fireproof` again a proper package and not just re-exports
of `@fireproof/core`. `use-fireproof` only depends on React and `@fireproof/core`
and hence is a good example of how to use `@fireproof/core`.

This also makes `@fireprof/core` free of React, which is good for folks
that use other frameworks.